feat: Updated 'collectAnswers' function to allow for optional 'message' parameter
The function 'collectAnswers' in 'quizUtils.js' was modified to include an extra optional parameter called 'message'. This was done to allow for the editing of message components when a user answers a quiz question. If a 'message' is provided, the 'answerButtons' will be edited on it; if not, the interaction's reply will be edited as before.